fix: drop unnecessary int->uint cast on shift count
gosec G115 flagged `baseHours << uint(exp)` as a potential int→uint overflow.
The cast was leftover from pre-Go-1.13 style; Go 1.13+ accepts signed integer
shift counts, and `exp` is already guarded non-negative (≤ 30) by the block
above. Removing the cast removes the finding without changing behavior.

+++ rss-worker/main.go
@@ -104,11 +104,12 @@ func nextCircuitOpenUntil(now time.Time, failures, threshold, baseHours, maxHour
 	}
 	exp := failures - threshold
 	// Cap the shift so `baseHours << exp` can't overflow int.
+	// (`exp` is already non-negative because of the `failures < threshold` guard above.)
 	const maxShift = 30
 	if exp > maxShift {
 		exp = maxShift
 	}
-	delay := baseHours << uint(exp)
+	delay := baseHours << exp
 	if maxHours > 0 && delay > maxHours {
 		delay = maxHours
 	}
